auth sevice integration test
