+++ auth-service/src/app.test.js
@@ -1,0 +1,40 @@
+const request = require('supertest');
+const app = require('./app'); 
+
+// Increase timeout to handle longer requests
+jest.setTimeout(10000); // 10 seconds
+
+describe('API Tests', () => {
+  // Test for customer login route
+  test('should return 200 for the customer login route', async () => {
+    const res = await request(app)
+      .post('/customer/login')
+      .send({
+        email: 'john.doe@example.com',  // Adjust with valid test data
+        password: 'himindu123'  // Adjust with valid test data
+      });
+
+    expect(res.statusCode).toBe(200);
+    expect(res.body).toHaveProperty('accessToken');
+  });
+
+  // Test for employee login route
+  test('should return 200 for the employee login route', async () => {
+    const res = await request(app)
+      .post('/employee/login')
+      .send({
+        email: 'john.doe@example.com',  // Adjust with valid test data
+        password: 'himindu123'      // Adjust with valid test data
+      });
+
+    expect(res.statusCode).toBe(200);
+    expect(res.body);
+  });
+});
+
+// Cleanup after tests if necessary
+afterAll(async () => {
+  // If you have database connections or other resources, clean them up
+  // Example:
+  // await db.end(); // Ensure this is the actual method to close your database connection
+});
