+++ auth-service/src/config/db.js
@@ -6,6 +6,7 @@ const db = mysql.createConnection({
     password: process.env.DB_PASSWORD || 'password',
     database: process.env.DB_NAME || 'pointmaster',
     port: process.env.DB_PORT || 3308,
+    charset: 'utf8mb4'
 });
 
 db.connect((err) => {
