+++ auth-service/src/package.json
@@ -4,7 +4,9 @@
   "description": "all authentication parts are done here",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "jest",
+    "test-coverage": "jest --coverage",
+    "test-watch": "jest --watchAll",
     "start": "node server.js",
     "dev": "nodemon server.js"
   },
@@ -22,5 +24,9 @@
     "nodemon": "^3.1.4",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0",
+    "supertest": "^7.0.0"
   }
 }
